requirements.txt
# Coalesce Migration Toolkit Dependencies

# Core HTTP requests for Coalesce API
requests>=2.28.0

# Data analysis and UUID lookup tables
pandas>=1.5.0

# Environment variable management
python-dotenv>=0.19.0

# Optional: Enhanced JSON handling (if needed)
# jsonschema>=4.0.0

# Optional: Progress bars for long operations (if desired)
# tqdm>=4.64.0

# Optional: Enhanced logging (if needed)
# colorlog>=6.6.0

# Development dependencies (uncomment if needed)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0